update customer page fully functioning and added confirmations on delete customer

+++ src/controller/customerGUIController.java
@@ -1,6 +1,5 @@
 package controller;
 
-import DBAccess.DBCountries;
 import DBAccess.DBCustomer;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
@@ -11,14 +10,12 @@ import javafx.scene.Scene;
 import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.stage.Stage;
-import model.Countries;
 import model.Customer;
 import utils.DBConnection;
 
 import java.io.IOException;
 import java.net.URL;
 import java.sql.PreparedStatement;
-import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.Objects;
 import java.util.Optional;
@@ -33,6 +30,8 @@ public class customerGUIController implements Initializable {
     public TableColumn<Customer, String> postalCol;
     public TableColumn<Customer, String> phoneCol;
     public TableColumn<Customer, String> locationCol;
+    private static Customer customerToModify;
+    private static int customerToModifyIndex;
 
     @Override
     public void initialize(URL url, ResourceBundle resourceBundle) {
@@ -41,6 +40,7 @@ public class customerGUIController implements Initializable {
     }
 
     public void backToMain(ActionEvent actionEvent) throws IOException {
+
         Alert alert = new Alert(Alert.AlertType.CONFIRMATION, "This will return to the Main Menu. Do you want to continue?");
         Optional<ButtonType> result = alert.showAndWait();
         if(result.isPresent() && result.get() == ButtonType.OK) {
@@ -56,6 +56,7 @@ public class customerGUIController implements Initializable {
     }
 
     public void addCustomer(ActionEvent actionEvent) throws IOException {
+
         Parent root = FXMLLoader.load(Objects.requireNonNull(getClass().getResource("/view/newCustomer.fxml")));
         Stage stage = (Stage) ((Button) (actionEvent.getSource())).getScene().getWindow();
 
@@ -67,31 +68,51 @@ public class customerGUIController implements Initializable {
     }
 
     public void modCustomer(ActionEvent actionEvent) throws IOException {
-        Parent root = FXMLLoader.load(Objects.requireNonNull(getClass().getResource("/view/modCustomer.fxml")));
-        Stage stage = (Stage) ((Button) (actionEvent.getSource())).getScene().getWindow();
+        customerToModify = customerList.getSelectionModel().getSelectedItem();
 
-        Scene scene = new Scene(root, 1000, 700);
-        stage.setTitle("Modify Customer");
-        stage.setScene(scene);
+        if(customerToModify == null) {
 
-        stage.show();
+            Alert alert = new Alert(Alert.AlertType.ERROR);
+            alert.setTitle("Error Dialog");
+            alert.setContentText("There are no products selected.");
+            alert.showAndWait();
+        } else{
+
+            customerToModify = customerList.getSelectionModel().getSelectedItem();
+            customerToModifyIndex = DBCustomer.getAllCustomers().indexOf(customerToModify);
+            Parent root = FXMLLoader.load(Objects.requireNonNull(getClass().getResource("/view/modCustomer.fxml")));
+            Stage stage = (Stage)((Button)(actionEvent.getSource())).getScene().getWindow();
+
+            Scene scene = new Scene(root,1000,700);
+            stage.setTitle("Update Customer");
+
+            stage.setScene(scene);
+            stage.show();
+        }
     }
 
     public void deleteCustomer() {
 
         Customer customerToDelete = customerList.getSelectionModel().getSelectedItem();
         if(customerToDelete == null) {
+
             Alert alert = new Alert(Alert.AlertType.ERROR);
             alert.setTitle("Error Dialog");
             alert.setContentText("There are no customers selected.");
             alert.showAndWait();
         } else {
             try {
-                String sql = "DELETE FROM customers WHERE Customer_ID = " + customerToDelete.getId();
 
-                PreparedStatement ps = DBConnection.getConnection().prepareStatement(sql);
-                ps.executeUpdate();
+                Alert alert = new Alert(Alert.AlertType.CONFIRMATION, "Would you like to permanently delete this customer?");
+                Optional<ButtonType> result = alert.showAndWait();
+                if(result.isPresent() && result.get() == ButtonType.OK) {
+                    String sql = "DELETE FROM customers WHERE Customer_ID = " + customerToDelete.getId();
+
+                    PreparedStatement ps = DBConnection.getConnection().prepareStatement(sql);
+                    ps.executeUpdate();
+                }
             } catch (NullPointerException | SQLException e) {
+
                 e.printStackTrace();
             }
 
@@ -111,5 +132,13 @@ public class customerGUIController implements Initializable {
         postalCol.setCellValueFactory(new PropertyValueFactory<>("postalCode"));
         phoneCol.setCellValueFactory(new PropertyValueFactory<>("phone"));
         locationCol.setCellValueFactory(new PropertyValueFactory<>("country"));
+    }
+
+    public static Customer getCustomerToModify() {
+        return customerToModify;
+    }
+
+    public static int getCustomerToModifyIndex() {
+        return customerToModifyIndex;
     }
 }
